requirements.txt
camelot-py==0.10.1
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.34.4
ghostscript==0.7
kiwisolver==1.4.4
numpy==1.23.1
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pdfminer.six==20220524
pdftotext==2.2.2
Pillow==9.2.0
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
PyPDF2==2.7.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
tabulate==0.8.10
